third_party/WebKit/LayoutTests/media/audio-controls-captions.html

<!DOCTYPE html>
<title>Tests that the closed captions button is not visible.</title>
<script src="media-controls.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<audio controls>
    <track src="track/captions-webvtt/captions-fast.vtt" default>
</audio>
<script>
async_test(function(t) {
    var audio = document.querySelector('audio');
    audio.src = 'content/test.oga';
    audio.onloadedmetadata = t.step_func_done(function() {
        assert_false(isClosedCaptionsButtonVisible(audio));
    });
});
</script>
